move getFailingPages into check.js, only return failed pages, reset failedContrast when swithing back and forth between colors
